my_top/21: add router for Menu move menu levels to helpers.tsx rename courses to [alias].tsx
