+++ layout/Menu/Menu.module.sass
@@ -35,7 +35,11 @@
 
   border-left: 1px solid #DFDFDF
 
+.secondLevelBlock
+  display: none
+
 .secondLevelBlockOpened
+  display: block
   &:last-child
     margin-bottom: 20px
